allow setting custom fields when create account
